refactor: 폴더 설정 버튼 컴포넌트 분리
