+++ src/pages/Folder/components/InfoContainer/InfoContainer.style.js
