+++ src/pages/Folder/components/InfoContainer/index.js
@@ -1,0 +1,1 @@
+export { default } from './InfoContainer';
